Module rome_rowan::syntax

Structs

Preorder

PreorderWithTokens

SendNode

Language-agnostic representation of the root node of a syntax tree, can be sent or shared between threads

SyntaxElementChildren

SyntaxElementKey

SyntaxList

A list of SyntaxNodes and/or SyntaxTokens

SyntaxNode

SyntaxNodeChildren

SyntaxToken

SyntaxTrivia

SyntaxTriviaPiece

SyntaxTriviaPiece gives access to the most granular information about the trivia that was specified by the lexer at the token creation time.

SyntaxTriviaPieceComments

SyntaxTriviaPieceNewline

SyntaxTriviaPieceSkipped

SyntaxTriviaPieceWhitespace

SyntaxTriviaPiecesIterator

TriviaPiece

Enums

SyntaxSlot

Each node has a slot for each of its children regardless if the child is present or not. A child that isn’t present either because it’s optional or because of a syntax error is stored in an SyntaxSlot::Empty to preserve the index of each child.

TriviaPieceKind

Traits

Language

SyntaxKind

Type tag for each node or token of a language

SyntaxNodeOptionExt

Trait with extension methods for Option.

Type Definitions

SyntaxElement