Trait rome_rowan::syntax::SyntaxKind

pub trait SyntaxKind: Debug + PartialEq + Copy {
    fn is_unknown(&self) -> bool;
    fn to_unknown(&self) -> Self;
    fn to_raw(&self) -> RawSyntaxKind;
    fn from_raw(raw: RawSyntaxKind) -> Self;
    fn is_root(&self) -> bool;
    fn is_list(&self) -> bool;
}

Type tag for each node or token of a language

Required Methods

fn is_unknown(&self) -> bool

Returns true if this is an unknown node kind.

fn to_unknown(&self) -> Self

Converts this into to the best matching unknown node kind.

fn to_raw(&self) -> RawSyntaxKind

Converts this kind to a raw syntax kind.

fn from_raw(raw: RawSyntaxKind) -> Self

Creates a syntax kind from a raw kind.

fn is_root(&self) -> bool

Returns true if this kind is for a root node.

fn is_list(&self) -> bool

Returns true if this kind is a list node.