Trait rome_rowan::AstNode

pub trait AstNode: Clone {
    type Language: Language;

    const KIND_SET: SyntaxKindSet<Self::Language>;

    fn can_cast(kind: <Self::Language as Language>::Kind) -> bool;
    fn cast(syntax: SyntaxNode<Self::Language>) -> Option<Self>
    where
        Self: Sized;
    fn syntax(&self) -> &SyntaxNode<Self::Language>;
    fn into_syntax(self) -> SyntaxNode<Self::Language>;

    fn cast_ref(syntax: &SyntaxNode<Self::Language>) -> Option<Self>
    where
        Self: Sized,
    { ... }
    fn try_cast(
        syntax: SyntaxNode<Self::Language>
    ) -> Result<Self, SyntaxNode<Self::Language>> { ... }
    fn try_cast_node<T: AstNode<Language = Self::Language>>(
        node: T
    ) -> Result<Self, T> { ... }
    fn unwrap_cast(syntax: SyntaxNode<Self::Language>) -> Self
    where
        Self: Sized,
    { ... }
    fn text(&self) -> String { ... }
    fn range(&self) -> TextRange { ... }
    fn clone_subtree(&self) -> Self
    where
        Self: Sized,
    { ... }
    fn parent<T: AstNode<Language = Self::Language>>(&self) -> Option<T> { ... }
}

The main trait to go from untyped SyntaxNode to a typed ast. The conversion itself has zero runtime cost: ast and syntax nodes have exactly the same representation: a pointer to the tree root and a pointer to the node itself.

Required Associated Types

type Language: Language

Required Associated Constants

const KIND_SET: SyntaxKindSet<Self::Language>

Required Methods

fn can_cast(kind: <Self::Language as Language>::Kind) -> bool

Returns true if a node with the given kind can be cased to this AST node.

fn cast(syntax: SyntaxNode<Self::Language>) -> Option<Self> where
Self: Sized,

Tries to cast the passed syntax node to this AST node.

Returns

None if the passed node is of a different kind. Some otherwise.

fn syntax(&self) -> &SyntaxNode<Self::Language>

Returns the underlying syntax node.

fn into_syntax(self) -> SyntaxNode<Self::Language>

Returns the underlying syntax node.

Provided Methods

fn cast_ref(syntax: &SyntaxNode<Self::Language>) -> Option<Self> where
Self: Sized,

Takes a reference of a syntax node and tries to cast it to this AST node.

Only creates a clone of the syntax node if casting the node is possible.

fn try_cast(
syntax: SyntaxNode<Self::Language>
) -> Result<Self, SyntaxNode<Self::Language>>

Tries to cast the passed syntax node to this AST node.

Returns

Ok if the passed node can be cast into this AstNode

Err(syntax) If the node is of another kind.

Examples

let mut builder = RawSyntaxTreeBuilder::new();

builder.start_node(RawLanguageKind::ROOT);
builder.start_node(RawLanguageKind::LITERAL_EXPRESSION);
builder.token(RawLanguageKind::STRING_TOKEN, "'abcd'");
builder.finish_node();
builder.finish_node();

let root_syntax = builder.finish();
let root = RawLanguageRoot::cast(root_syntax.clone()).expect("Root to be a raw language root");

// Returns `OK` because syntax is a `RawLanguageRoot`
assert_eq!(RawLanguageRoot::try_cast(root.syntax().clone()), Ok(root.clone()));
// Returns `Err` with the syntax node passed to `try_cast` because `root` isn't a `LiteralExpression`
assert_eq!(LiteralExpression::try_cast(root.syntax().clone()), Err(root_syntax));

fn try_cast_node<T: AstNode<Language = Self::Language>>(
node: T
) -> Result<Self, T>

Tries to cast the AST node into this node.

Returns

Ok if the passed node can be cast into this AstNode

Err if the node is of another kind

let mut builder = RawSyntaxTreeBuilder::new();

builder.start_node(RawLanguageKind::ROOT);
builder.start_node(RawLanguageKind::LITERAL_EXPRESSION);
builder.token(RawLanguageKind::STRING_TOKEN, "'abcd'");
builder.finish_node();
builder.finish_node();

let root_syntax = builder.finish();
let root = RawLanguageRoot::cast(root_syntax.clone()).expect("Root to be a raw language root");

// Returns `OK` because syntax is a `RawLanguageRoot`
assert_eq!(RawLanguageRoot::try_cast_node(root.clone()), Ok(root.clone()));

// Returns `Err` with the node passed to `try_cast_node` because `root` isn't a `LiteralExpression`
assert_eq!(LiteralExpression::try_cast_node(root.clone()), Err(root.clone()));

fn unwrap_cast(syntax: SyntaxNode<Self::Language>) -> Self where
Self: Sized,

Cast this node to this AST node

Panics

Panics if the underlying node cannot be cast to this AST node

fn text(&self) -> String

Returns the string representation of this node without the leading and trailing trivia

fn range(&self) -> TextRange

fn clone_subtree(&self) -> Self where
Self: Sized,

fn parent<T: AstNode<Language = Self::Language>>(&self) -> Option<T>