Trait rome_rowan::AstNodeExt

pub trait AstNodeExt: AstNode {
    fn replace_node_discard_trivia<N>(
        self,
        prev_node: N,
        next_node: N
    ) -> Option<Self>
    where
        N: AstNode<Language = Self::Language>,
        Self: Sized;
    fn replace_node<N>(self, prev_node: N, next_node: N) -> Option<Self>
    where
        N: AstNode<Language = Self::Language>,
        Self: Sized;
    fn replace_token_discard_trivia(
        self,
        prev_token: SyntaxToken<Self::Language>,
        next_token: SyntaxToken<Self::Language>
    ) -> Option<Self>
    where
        Self: Sized;
    fn replace_token(
        self,
        prev_token: SyntaxToken<Self::Language>,
        next_token: SyntaxToken<Self::Language>
    ) -> Option<Self>
    where
        Self: Sized;
    fn detach(self) -> Self;
}

Required Methods

fn replace_node_discard_trivia<N>(
self,
prev_node: N,
next_node: N
) -> Option<Self> where
N: AstNode<Language = Self::Language>,
Self: Sized,

Return a new version of this node with the node prev_node replaced with next_node

prev_node can be a direct child of this node, or an indirect child through any descendant node

Returns None if prev_node is not a descendant of this node

fn replace_node<N>(self, prev_node: N, next_node: N) -> Option<Self> where
N: AstNode<Language = Self::Language>,
Self: Sized,

Return a new version of this node with the node prev_node replaced with next_node, transfering the leading and trailing trivia of prev_node to next_node

prev_node can be a direct child of this node, or an indirect child through any descendant node

Returns None if prev_node is not a descendant of this node

fn replace_token_discard_trivia(
self,
prev_token: SyntaxToken<Self::Language>,
next_token: SyntaxToken<Self::Language>
) -> Option<Self> where
Self: Sized,

Return a new version of this node with the token prev_token replaced with next_token

prev_token can be a direct child of this node, or an indirect child through any descendant node

Returns None if prev_token is not a descendant of this node

fn replace_token(
self,
prev_token: SyntaxToken<Self::Language>,
next_token: SyntaxToken<Self::Language>
) -> Option<Self> where
Self: Sized,

Return a new version of this node with the token prev_token replaced with next_token, transfering the leading and trailing trivia of prev_token to next_token

prev_token can be a direct child of this node, or an indirect child through any descendant node

Returns None if prev_token is not a descendant of this node

fn detach(self) -> Self