Trait rome_rowan::AstSeparatedList

pub trait AstSeparatedList {
    type Language: Language;
    type Node: AstNode<Language = Self::Language>;

    fn syntax_list(&self) -> &SyntaxList<Self::Language>;
    fn into_syntax_list(self) -> SyntaxList<Self::Language>;

    fn elements(
        &self
    ) -> AstSeparatedListElementsIterator<Self::Language, Self::Node> { ... }
    fn separators(&self) -> AstSeparatorIterator<Self::Language, Self::Node> { ... }
    fn iter(&self) -> AstSeparatedListNodesIterator<Self::Language, Self::Node> { ... }
    fn first(&self) -> Option<SyntaxResult<Self::Node>> { ... }
    fn last(&self) -> Option<SyntaxResult<Self::Node>> { ... }
    fn is_empty(&self) -> bool { ... }
    fn len(&self) -> usize { ... }
    fn trailing_separator(&self) -> Option<SyntaxToken<Self::Language>> { ... }
}

List of nodes where every two nodes are separated by a token. For example, the elements of an array where every two elements are separated by a comma token. The list expects that the underlying syntax node has a slot for every node and separator even if they are missing from the source code. For example, a list for a b where the , separator is missing contains the slots Node(a), Empty, Node(b). This also applies for missing nodes: the list for , b, must have the slots Empty, Token(,), Node(b), Token(,).

Required Associated Types

type Language: Language

type Node: AstNode<Language = Self::Language>

Required Methods

fn syntax_list(&self) -> &SyntaxList<Self::Language>

Returns the underlying syntax list

fn into_syntax_list(self) -> SyntaxList<Self::Language>

Returns the underlying syntax list

Provided Methods

fn elements(
&self
) -> AstSeparatedListElementsIterator<Self::Language, Self::Node>

Returns an iterator over all nodes with their trailing separator

fn separators(&self) -> AstSeparatorIterator<Self::Language, Self::Node>

Returns an iterator over all separator tokens

fn iter(&self) -> AstSeparatedListNodesIterator<Self::Language, Self::Node>

Returns an iterator over all nodes

fn first(&self) -> Option<SyntaxResult<Self::Node>>

Returns the first node

fn last(&self) -> Option<SyntaxResult<Self::Node>>

Returns the last node

fn is_empty(&self) -> bool

fn len(&self) -> usize

fn trailing_separator(&self) -> Option<SyntaxToken<Self::Language>>