Trait rome_rowan::SyntaxFactory

pub trait SyntaxFactory: Debug {
    type Kind: SyntaxKind;

    fn make_syntax(
        kind: Self::Kind,
        children: ParsedChildren<'_, Self::Kind>
    ) -> RawSyntaxNode<Self::Kind>;

    fn make_node_list_syntax<F>(
        kind: Self::Kind,
        children: ParsedChildren<'_, Self::Kind>,
        can_cast: F
    ) -> RawSyntaxNode<Self::Kind>
    where
        F: Fn(Self::Kind) -> bool,
    { ... }
    fn make_separated_list_syntax<F>(
        kind: Self::Kind,
        children: ParsedChildren<'_, Self::Kind>,
        can_cast: F,
        separator: Self::Kind,
        allow_trailing: bool
    ) -> RawSyntaxNode<Self::Kind>
    where
        F: Fn(Self::Kind) -> bool,
    { ... }
}

Factory for creating syntax nodes of a particular kind.

Required Associated Types

type Kind: SyntaxKind

The syntax kind used by the nodes constructed by this syntax factory.

Required Methods

fn make_syntax(
kind: Self::Kind,
children: ParsedChildren<'_, Self::Kind>
) -> RawSyntaxNode<Self::Kind>

Creates a new syntax node of the passed kind with the given children.

The children contains the parsed direct children of the node. There may be fewer children in case there’s a syntax error and a required child or an optional child isn’t present in the source code. The make_syntax implementation must then fill in empty slots to match the slots as they’re defined in the grammar.

The implementation is free to change the kind of the node but that has the consequence that such a node will not be cached. The reason for not caching these nodes is that the cache lookup is performed before calling make_syntax, thus querying the cache with the old kind.

It’s important that the factory function is idempotent, meaning, calling the function multiple times with the same kind and children returns syntax nodes with the same structure. This is important because the returned nodes may be cached by kind and what children are present.

Provided Methods

fn make_node_list_syntax<F>(
kind: Self::Kind,
children: ParsedChildren<'_, Self::Kind>,
can_cast: F
) -> RawSyntaxNode<Self::Kind> where
F: Fn(Self::Kind) -> bool,

Crates a node list syntax node. Validates if all elements are valid and changes the node’s kind to SyntaxKind::to_unknown if that’s not the case.

fn make_separated_list_syntax<F>(
kind: Self::Kind,
children: ParsedChildren<'_, Self::Kind>,
can_cast: F,
separator: Self::Kind,
allow_trailing: bool
) -> RawSyntaxNode<Self::Kind> where
F: Fn(Self::Kind) -> bool,

Creates a separated list syntax node. Validates if the elements are valid, are correctly separated by the specified separator token.

It changes the kind of the node to SyntaxKind::to_unknown if an element isn’t a valid list-node nor separator.

It inserts empty slots for missing elements or missing markers