Trait rome_rowan::SyntaxRewriter

pub trait SyntaxRewriter {
    type Language: Language;

    fn transform(
        &mut self,
        node: SyntaxNode<Self::Language>
    ) -> SyntaxNode<Self::Language>
    where
        Self: Sized,
    { ... }
    fn visit_node(
        &mut self,
        node: SyntaxNode<Self::Language>
    ) -> VisitNodeSignal<Self::Language> { ... }
    fn visit_token(
        &mut self,
        token: SyntaxToken<Self::Language>
    ) -> SyntaxToken<Self::Language> { ... }
}

A visitor that re-writes a syntax tree while visiting the nodes.

The rewriter visits the nodes in pre-order from top-down. Meaning, it first visits the root, and then visits the children of the root from left to right, recursively traversing into child nodes and calling visit_node for every node.

Inspired by Roslyn’s CSharpSyntaxRewriter

Unsupported

The current implementation does not yet support node removal.

Examples

Implementation of a rewritter that replaces all literal expression nodes that contain a number token with an unknown node.

struct ReplaceNumberLiteralRewriter;

impl SyntaxRewriter for ReplaceNumberLiteralRewriter {
    type Language = RawLanguage;

    fn visit_node(
        &mut self,
        node: SyntaxNode<Self::Language>,
    ) -> VisitNodeSignal<Self::Language> {
        match node.kind() {
            RawLanguageKind::LITERAL_EXPRESSION => {
                let expression = LiteralExpression::unwrap_cast(node);

                let mut token = expression
                    .syntax()
                    .slots()
                    .nth(0)
                    .unwrap()
                    .into_token()
                    .unwrap();

                match token.kind() {
                    RawLanguageKind::NUMBER_TOKEN => {
                        // Use your language's syntax factory instead
                        let unknown_node = SyntaxNode::new_detached(
                            RawLanguageKind::UNKNOWN,
                            once(Some(token.into())),
                        );

                        VisitNodeSignal::Replace(unknown_node)
                    }
                    // Not interested in string literal expressions, continue traversal
                    _ => VisitNodeSignal::Traverse(expression.into_syntax()),
                }
            }
            _ => {
                // Traverse into the childrens of node
                VisitNodeSignal::Traverse(node)
            }
        }
    }
}

let mut builder = RawSyntaxTreeBuilder::new();

builder.start_node(RawLanguageKind::ROOT);
builder.start_node(RawLanguageKind::SEPARATED_EXPRESSION_LIST);

builder.start_node(RawLanguageKind::LITERAL_EXPRESSION);
builder.token(RawLanguageKind::NUMBER_TOKEN, "5");
builder.finish_node();

builder.start_node(RawLanguageKind::LITERAL_EXPRESSION);
builder.token(RawLanguageKind::STRING_TOKEN, "'abcd'");
builder.finish_node();

builder.finish_node();
builder.finish_node();

let root = builder.finish();

let transformed = ReplaceNumberLiteralRewriter.transform(root.clone());

let original_literal_expressions: Vec<_> = root
    .descendants()
    .filter(|p| p.kind() == RawLanguageKind::LITERAL_EXPRESSION)
    .collect();

assert_ne!(
    &root, &transformed,
    "It returns a new root with the updated children"
);

let literal_expressions: Vec<_> = transformed
    .descendants()
    .filter(|p| p.kind() == RawLanguageKind::LITERAL_EXPRESSION)
    .collect();

 // The literal expression containing a string token should be unchanged
 assert_eq!(&literal_expressions, &original_literal_expressions[1..]);

 let mut unknowns: Vec<_> = transformed
    .descendants()
    .filter(|p| p.kind() == RawLanguageKind::UNKNOWN)
    .collect();

// It replaced the number literal expression with an unknown node.
assert_eq!(unknowns.len(), 1);
assert_eq!(unknowns.pop().unwrap().text(), "5");

Required Associated Types

type Language: Language

Provided Methods

fn transform(
&mut self,
node: SyntaxNode<Self::Language>
) -> SyntaxNode<Self::Language> where
Self: Sized,

Recursively transforms the subtree of node by calling visit_node for every token and visit_token for every token in the subtree.

Returns a new syntax tree reflecting the changes by the rewriter if it replaced any node and returns node if no changes were made.

fn visit_node(
&mut self,
node: SyntaxNode<Self::Language>
) -> VisitNodeSignal<Self::Language>

Called for every node in the tree. The method should return a signal specifying what should be done with the node

VisitNodeSignal::Traverse: Recourse into node so that visit_node gets called for all children of node. The node will only be replaced if any node in its subtree changes.

VisitNodeSignal::Replace: Replaces node with the node specified in the Replace variant. It’s your responsibility to call traverse for any child of node for which you want the rewritter to recurse into its content.

fn visit_token(
&mut self,
token: SyntaxToken<Self::Language>
) -> SyntaxToken<Self::Language>

Called for every token in the tree. Returning a new token changes the token in the parent node.